requirements.txt
-i https://pypi.org/simple

# Common requirements
# -----------------------

setuptools
wheel>=0.35.1

# Project requirements
# -----------------------

isogeo-pysdk>=3.5.0
openpyxl>=3.0.5
python-dotenv

# Dev requirements
# -----------------------

black==20.8b1
doc8
flake8
pre-commit
pylint>=2.6.0
pytest<6.2
pytest-cov>=2.10.1
sphinx
